+++ lib/mix/lib/mix/tasks/test.ex
@@ -49,8 +49,42 @@ defmodule Mix.Tasks.Test do
   * `--no-compile` - do not compile, even if files require compilation
   * `--no-start` - do not start applications after compilation
   * `--no-color` - disable color in the output
-  * `--include` - skip tests that do not match the filter
-  * `--exclude` - skip tests that match the filter
+  * `--include` - include tests that match the filter
+  * `--exclude` - exclude tests that match the filter
+  * `--only` - run only tests that match the filter
+
+  ## Filters
+
+  ExUnit provides tags and filtering functionality that allows developers
+  to select which tests to run. The most common functionality is to exclude
+  some particular tests from running by default in your test helper file:
+
+      # Exclude all external tests from running
+      ExUnit.configure exclude: [external: true]
+
+  Then, whenever desired, those tests could be included in the run via the
+  `--include` flag:
+
+      mix test --include external:true
+
+  The example above will run all tests that have the external flag set to
+  true. It is also possible to include all examples that have a given tag,
+  regardless of its value:
+
+      mix test --include external
+
+  Note that all tests are included by default, so unless they are excluded
+  first (either in the test helper or via the `--exclude` option), the
+  `--include` flag has no effect.
+
+  For this reason, mix also provides an `--only` option that excludes all
+  tests and includes only the given ones:
+
+      mix test --only external
+
+  Which is equivalent to:
+
+      mix test --include external --exclude test
 
   ## Configuration
 
@@ -106,7 +140,7 @@ defmodule Mix.Tasks.Test do
     # test_helper so that the configuration is available in test_helper
     # Then configure exunit again it so command line options override
     # test_helper
-    opts = prepare_opts(opts)
+    opts = ex_unit_opts(opts)
     ExUnit.configure(opts)
 
     test_paths = project[:test_paths] || ["test"]
@@ -121,6 +155,42 @@ defmodule Mix.Tasks.Test do
     Kernel.ParallelRequire.files files
   end
 
+  @doc false
+  def ex_unit_opts(opts) do
+    opts = opts
+           |> filter_opts(:include)
+           |> filter_opts(:exclude)
+           |> filter_only_opts()
+
+    Dict.take(opts, [:trace, :max_cases, :color, :include, :exclude, :seed])
+  end
+
+  defp parse_filters(opts, key) do
+    if Keyword.has_key?(opts, key) do
+      ExUnit.Filters.parse(Keyword.get_values(opts, key))
+    end
+  end
+
+  defp filter_opts(opts, key) do
+    if filters = parse_filters(opts, key) do
+      Keyword.put(opts, key, filters)
+    else
+      opts
+    end
+  end
+
+  defp filter_only_opts(opts) do
+    if filters = parse_filters(opts, :only) do
+      opts
+      |> Keyword.put_new(:include, [])
+      |> Keyword.put_new(:exclude, [])
+      |> Keyword.update!(:include, &(filters ++ &1))
+      |> Keyword.update!(:exclude, &[:test|&1])
+    else
+      opts
+    end
+  end
+
   defp require_test_helper(dir) do
     file = Path.join(dir, "test_helper.exs")
 
@@ -129,19 +199,5 @@ defmodule Mix.Tasks.Test do
     else
       raise Mix.Error, message: "Cannot run tests because test helper file #{inspect file} does not exist"
     end
-  end
-
-  defp prepare_opts(opts) do
-    if opts[:include] do
-      inclusions = ExUnit.Filters.parse(Keyword.get_values(opts, :include))
-      opts = Keyword.put(opts, :include, inclusions)
-    end
-
-    if opts[:exclude] do
-      exclusions = ExUnit.Filters.parse(Keyword.get_values(opts, :exclude))
-      opts = Keyword.put(opts, :exclude, exclusions)
-    end
-
-    Dict.take(opts, [:trace, :max_cases, :color, :include, :exclude, :seed])
   end
 end